generation of the certificate with dummy data
